+++ content/src/logic/deployments.ts
@@ -1,13 +1,12 @@
 import { Entity } from '@dcl/schemas'
 import { ILoggerComponent } from '@well-known-components/interfaces'
-import SQL from 'sql-template-strings'
 import { FailedDeployment } from '../ports/failedDeploymentsCache'
 import { AuditInfo, Deployment } from '../service/deployments/types'
 import { DeploymentContext } from '../service/Service'
 import { deployEntityFromRemoteServer } from '../service/synchronization/deployRemoteEntity'
 import { IGNORING_FIX_ERROR } from '../service/validations/server'
-import { AppComponents } from '../types'
-import { deploymentExists, saveContentFiles, saveDeployment } from './database-queries/deployments-queries'
+import { AppComponents, DeploymentId } from '../types'
+import { calculateOverwrittenByMany1, calculateOverwrittenByMany2, calculateOverwrote, deploymentExists, saveContentFiles, saveDeployment } from './database-queries/deployments-queries'
 
 export async function isEntityDeployed(
   components: Pick<AppComponents, 'deployedEntitiesBloomFilter' | 'database' | 'metrics'>,
@@ -97,8 +96,6 @@ export function mapDeploymentsToEntities(deployments: Deployment[]): Entity[] {
 }
 
 
-type DeploymentId = number
-
 export async function saveDeploymentAndContentFiles(
   components: Pick<AppComponents, 'database'>,
   entity: Entity,
@@ -116,45 +113,13 @@ export async function calculateOverwrites(
   components: Pick<AppComponents, 'database' | 'repository'>,
   entity: Entity
 ): Promise<{ overwrote: Set<DeploymentId>; overwrittenBy: DeploymentId | null }> {
-  const overwrote: DeploymentId[] = (await components.database.queryWithValues<{ id: number }>(
-    SQL`
-          SELECT dep1.id
-          FROM deployments AS dep1
-          LEFT JOIN deployments AS dep2 ON dep1.deleter_deployment = dep2.id
-          WHERE dep1.entity_type = ${entity.type} AND
-              dep1.entity_pointers && ${entity.pointers} AND
-              (dep1.entity_timestamp < to_timestamp(${entity.timestamp} / 1000.0) OR (dep1.entity_timestamp = to_timestamp(${entity.timestamp} / 1000.0) AND dep1.entity_id < ${entity.id})) AND
-              (dep2.id IS NULL OR dep2.entity_timestamp > to_timestamp(${entity.timestamp} / 1000.0) OR (dep2.entity_timestamp = to_timestamp(${entity.timestamp} / 1000.0) AND dep2.entity_id > ${entity.id}))
-          ORDER BY dep1.entity_timestamp DESC, dep1.entity_id DESC`
-  )).rows.map((row) => row.id)
+  const overwrote = await calculateOverwrote(components, entity)
 
-  const q = SQL`
-  SELECT deployments.id
-  FROM active_pointers as ap
-           INNER JOIN deployments on ap.entity_id = deployments.entity_id
-  WHERE ap.pointer IN (`
-  const pointers = Array.from(entity.pointers)
-    .map((pointer, idx) => (idx < entity.pointers.length - 1) ? SQL`${pointer},` : SQL`${pointer}`)
-  pointers.forEach((pointer) => q.append(pointer))
-  q.append(SQL`)
-          AND deployments.entity_type = ${entity.type}
-          AND (deployments.entity_timestamp > to_timestamp(${entity.timestamp} / 1000.0) OR (deployments.entity_timestamp = to_timestamp(${entity.timestamp} / 1000.0) AND deployments.entity_id > ${entity.id}))
-        ORDER BY deployments.entity_timestamp, deployments.entity_id
-        LIMIT 1`)
-  let overwrittenByMany = (await components.database.queryWithValues<{ id: number }>(q)).rows
+  let overwrittenByMany = await calculateOverwrittenByMany1(components, entity)
 
   if (overwrittenByMany.length === 0 && entity.type === 'scene') {
     // Scene overwrite determination can be tricky. If none was detected use this other query (slower but safer)
-    overwrittenByMany = (await components.database.queryWithValues<{ id: number }>(
-      SQL`
-      SELECT deployments.id
-      FROM deployments
-      WHERE deployments.entity_type = ${entity.type} AND
-          deployments.entity_pointers && ${entity.pointers} AND
-          (deployments.entity_timestamp > to_timestamp(${entity.timestamp} / 1000.0) OR (deployments.entity_timestamp = to_timestamp(${entity.timestamp} / 1000.0) AND deployments.entity_id > ${entity.id}))
-      ORDER BY deployments.entity_timestamp, deployments.entity_id
-      LIMIT 1`
-    )).rows
+    overwrittenByMany = await calculateOverwrittenByMany2(components, entity)
   }
 
   let overwrittenBy: DeploymentId | null = null
